requirements.txt
Django==3.2.16
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
djoser==2.1.0
debugpy==1.6.4